sumo código sin memoria dinámica
